+++ rdkit_utilities/rdDistGeom.py
@@ -229,7 +229,7 @@ def SelectDiverseConformers(
     """
     from .rdMolAlign import GetBestConformerRMS
     from .rdmolops import KeepConformerIds
-    from ._utils import get_maximally_diverse_indices
+    from .utils import get_maximally_diverse_indices
 
     rms_matrix = GetBestConformerRMS(mol, heavyAtomsOnly=True)
     diverse_indices = get_maximally_diverse_indices(
@@ -341,7 +341,7 @@ def CalculateElectrostaticEnergy(
     energies: numpy.ndarray
         Array of energies, one for each conformer
     """
-    from ._utils import compute_atom_distance_matrix
+    from .utils import compute_atom_distance_matrix
 
     conformers = np.array([c.GetPositions() for c in mol.GetConformers()])
     distances = compute_atom_distance_matrix(conformers)
