+++ rdkit_utilities/tests/test_utils.py
@@ -5,9 +5,9 @@ from numpy.testing import assert_allclose
 
 
 def test_compute_distance_matrix():
-    from rdkit_utilities import _utils
+    from rdkit_utilities import utils
     arr = np.arange(24).reshape((2, 4, 3))
-    rms = _utils.compute_atom_distance_matrix(arr)
+    rms = utils.compute_atom_distance_matrix(arr)
     assert rms.shape == (2, 4, 4)
 
     first_row = [0, 5.19615242, 10.39230485, 15.58845727]
@@ -27,14 +27,14 @@ def test_compute_distance_matrix():
     (5.4, 4, [0, 3, 1]),
 ])
 def test_get_maximally_diverse_indices(threshold, n_indices, indices):
-    from rdkit_utilities import _utils
+    from rdkit_utilities import utils
     arr = np.array([
         [ 0.         , 5.49615242 ,10.39230485 ,15.58845727],
         [ 5.49615242 , 0.         , 5.19615242 ,10.39230485],
         [10.39230485 , 5.19615242 , 0.         , 5.19615242],
         [15.58845727 ,10.39230485 , 5.19615242 , 0.        ],
     ])
-    computed = _utils.get_maximally_diverse_indices(
+    computed = utils.get_maximally_diverse_indices(
         arr,
         distance_threshold=threshold,
         n_indices=n_indices,
